+++ src/Anthropic.Tests/Models/Beta/Models/ModelListParamsTest.cs
@@ -90,7 +90,7 @@ public class ModelListParamsTest : TestBase
 
         Assert.Equal(
             new Uri(
-                "https://api.anthropic.com/v1/models?after_id=after_id&before_id=before_id&limit=1"
+                "https://api.anthropic.com/v1/models?beta=true&after_id=after_id&before_id=before_id&limit=1"
             ),
             url
         );