+++ src/Anthropic/Models/Messages/Batches/BatchResultsParams.cs
@@ -111,6 +111,7 @@ public record class BatchResultsParams : ParamsBase
     internal override void AddHeadersToRequest(HttpRequestMessage request, ClientOptions options)
     {
         ParamsBase.AddDefaultHeaders(request, options);
+        request.Headers.Add("Accept", "application/x-jsonl");
         foreach (var item in this.RawHeaderData)
         {
             ParamsBase.AddHeaderElementToRequest(request, item.Key, item.Value);