+++ src/Anthropic.Vertex/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.1.0 (2026-03-15)
+
+Full Changelog: [Vertex-v0.0.1...Vertex-v0.1.0](https://github.com/anthropics/anthropic-sdk-csharp/compare/Vertex-v0.0.1...Vertex-v0.1.0)
+
+### Features
+
+* **tests:** update mock server ([775f7d1](https://github.com/anthropics/anthropic-sdk-csharp/commit/775f7d174fe5729675c7fe91d1c7bd9749e7c053))
+
+
+### Chores
+
+* **client:** update microsoft.bcl.memory ([5fcec90](https://github.com/anthropics/anthropic-sdk-csharp/commit/5fcec90dffb11bb3f99ff5ecd919d193c8994abb))
+* **docs:** add undocumented parameters to readme ([1d996bb](https://github.com/anthropics/anthropic-sdk-csharp/commit/1d996bb26dc18826832dc56ed44fb82669f1ee68))
+* **docs:** minor example cleanup ([e14da75](https://github.com/anthropics/anthropic-sdk-csharp/commit/e14da75bf138420ed889c2fbc0b6f80956345e4d))
+
 ## 0.0.1 (2026-02-12)
 
 Full Changelog: [...Vertex-v0.0.1](https://github.com/anthropics/anthropic-sdk-csharp/compare/...Vertex-v0.0.1)